pyproject.toml
[project]
name = "youtube-playlist-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.192.0",
    "mutagen>=1.47.0",
    "pyfiglet>=1.0.4",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "selenium>=4.41.0",
    "setuptools>=82.0.1",
    "textual>=8.1.1",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
    "yt-dlp>=2026.3.3",
]
